requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-extensions==3.2.3
django-unfold==0.47.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
